New commit

+++ src/main/java/com/example/springdemo/model/User.java
@@ -12,7 +12,7 @@ public class User {
 	@Id
 	@Column(name = "id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	
 	@Column(nullable = false, unique = true, length = 45)  //"'", "\"", "--", "/", "\\", "&", ";","||
 	//@Pattern(regexp="^[a-z0-9_A-ZĚŠČŘŽÝÁÍÉěščřžýáíéúůÚŮ]{3,10}$", message="INVALID INPUT")
@@ -112,11 +112,11 @@ public class User {
 		this.failedAttempt = failedAttempt;
 	}
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
