+++ src/main/java/com/example/springdemo/repository/UserRepository.java
@@ -10,7 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.Optional;
 
 @Transactional
-public interface UserRepository extends JpaRepository<User, Long>, CustomUserRepository, CrudRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer>, CustomUserRepository, CrudRepository<User, Integer> {
 	@Query("SELECT u FROM User u WHERE u.username = ?1")
 	public User findByUsername(String username);
